feat(odyssey-react-mui): create hexToRgb util
